requirements.txt
# Веб-фреймворк
Django>=5.0,<5.1

# База данных (если используешь PostgreSQL в продакшене — можно заменить)
psycopg2-binary>=2.9; python_version >= '3.8'
# Или для SQLite (встроенный, подойдёт для старта):
# (ничего не нужно — встроен в Python)

# Для REST API (если планируешь)
djangorestframework>=3.15

# Анализ сигналов и данных
numpy>=1.26
scipy>=1.13
pandas>=2.2

# Визуализация (если графики генерируются на сервере)
# matplotlib>=3.8

# WSGI/ASGI сервер для продакшена
gunicorn>=22.0
uvicorn[standard]>=0.30  # если будешь использовать ASGI (например, с FastAPI-совместимостью или Channels)

# Утилиты
python-dotenv>=1.0         # для загрузки .env
requests>=2.31             # HTTP-запросы (часто нужно)

asyncpg==0.30.0         #асинхронность
Flask==3.1.2            #веб тест
Flask-SQLAlchemy==3.1.1 #веб тест